LNP-1607: ♻️  🚧  split the cache warmer into a base class and two subclasses, one for each front end.
